# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pet Hotel & Grooming System API",
    "description": "Contract-first OpenAPI for the Pet Hotel & Grooming System.\n\n**S1 status:** endpoints are stubs (HTTP 501) with complete request/response schemas — this document is the single source of truth for backend + frontend. Business logic, state machines and real payment/notification land in S2+.\n\nAuthoritative data model: `diagrams/puml/class_diagram_plantuml.puml`. Internal modules (payment, notification, audit) are not exposed as endpoints.",
    "version": "0.1.0-S1"
  },
  "paths": {
...
    "/api/v1/bookings/services": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "List bookable service items — Standard/Deluxe room, Basic/Full grooming (FR-03.1/03.7)",
        "operationId": "list_services_api_v1_bookings_services_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ServiceCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceItemOut"
                  },
                  "title": "Response List Services Api V1 Bookings Services Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/availability": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Check availability for a service + window (FR-03.2/06.2)",
        "operationId": "check_availability_api_v1_bookings_availability_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Item Id"
            }
          },
          {
            "name": "start_at",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start At"
            }
          },
          {
            "name": "end_at",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/pending-review": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "List PendingReview bookings — staff (FR-02.8, seq5)",
        "operationId": "list_pending_review_api_v1_bookings_pending_review_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "type": "array",
                  "title": "Response List Pending Review Api V1 Bookings Pending Review Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/bookings": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Create a lodging/grooming booking (FR-03.1, seq1)",
        "operationId": "create_booking_api_v1_bookings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "409": {
            "description": "Rejected — pet danger level is High (FR-02.8)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "List my bookings — current + history (FR-03.4)",
        "operationId": "list_bookings_api_v1_bookings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BookingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "title": "Response List Bookings Api V1 Bookings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/{booking_id}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Get booking detail incl. items + pets (FR-03.4/04.4)",
        "operationId": "get_booking_api_v1_bookings__booking_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/{booking_id}/deposit": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Pay deposit (30%) and confirm (FR-05.3, seq1)",
        "operationId": "pay_deposit_api_v1_bookings__booking_id__deposit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositPaymentIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingPaymentResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/{booking_id}/final-payment": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Settle the final balance (FR-05.3)",
        "operationId": "pay_final_api_v1_bookings__booking_id__final_payment_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalPaymentIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingPaymentResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/{booking_id}/review": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Approve/Reject a PendingReview booking — staff (FR-02.8, seq5)",
        "operationId": "review_booking_api_v1_bookings__booking_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cancellation/bookings/{booking_id}": {
      "post": {
        "tags": [
          "Cancellation"
        ],
        "summary": "Cancel a booking + apply 24h refund rule (FR-03.3/05.4, seq4)",
        "operationId": "cancel_booking_api_v1_cancellation_bookings__booking_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancellationResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailabilityOut": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/AvailabilitySlotOut"
            },
            "type": "array",
            "title": "Slots"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "AvailabilityOut",
        "description": "FR-03.2 / FR-06.2 — only currently-available slots are returned."
      },
      "AvailabilitySlotOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "serviceItemId": {
            "type": "string",
            "format": "uuid",
            "title": "Serviceitemid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startat"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "title": "Endat"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "serviceItemId",
          "startAt",
          "endAt",
          "capacity"
        ],
        "title": "AvailabilitySlotOut"
      },
...
      "BookedPetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "petId": {
            "type": "string",
            "format": "uuid",
            "title": "Petid"
          },
          "kennelId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kennelid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "petId"
        ],
        "title": "BookedPetOut"
      },
      "BookingCreateIn": {
        "properties": {
          "startAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startat"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "title": "Endat"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookingItemIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "startAt",
          "endAt",
          "items"
        ],
        "title": "BookingCreateIn",
        "description": "FR-03.1 — choose pet(s), service(s) and a time window. Danger gating\n(FR-02.8) decides the resulting status (PendingDeposit / PendingReview) or a\n409 rejection when the pet's danger level is High."
      },
      "BookingDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "title": "Ownerid"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startat"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "title": "Endat"
          },
          "totalAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "depositAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "cancelledAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelledat"
          },
          "cancelReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelreason"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookingItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "bookedPets": {
            "items": {
              "$ref": "#/components/schemas/BookedPetOut"
            },
            "type": "array",
            "title": "Bookedpets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "status",
          "startAt",
          "endAt",
          "totalAmount",
          "depositAmount",
          "createdAt"
        ],
        "title": "BookingDetailOut"
      },
      "BookingItemIn": {
        "properties": {
          "serviceItemId": {
            "type": "string",
            "format": "uuid",
            "title": "Serviceitemid"
          },
          "petId": {
            "type": "string",
            "format": "uuid",
            "title": "Petid"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "serviceItemId",
          "petId"
        ],
        "title": "BookingItemIn"
      },
      "BookingItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "serviceItemId": {
            "type": "string",
            "format": "uuid",
            "title": "Serviceitemid"
          },
          "bookedPetId": {
            "type": "string",
            "format": "uuid",
            "title": "Bookedpetid"
          },
          "unitPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "serviceItemId",
          "bookedPetId",
          "unitPrice",
          "quantity"
        ],
        "title": "BookingItemOut"
      },
      "BookingOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "title": "Ownerid"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startat"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "title": "Endat"
          },
          "totalAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "depositAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "cancelledAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelledat"
          },
          "cancelReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelreason"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "status",
          "startAt",
          "endAt",
          "totalAmount",
          "depositAmount",
          "createdAt"
        ],
        "title": "BookingOut"
      },
      "BookingPaymentResultOut": {
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/BookingOut"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentResultOut"
          }
        },
        "type": "object",
        "required": [
          "booking",
          "payment"
        ],
        "title": "BookingPaymentResultOut"
      },
      "BookingStatus": {
        "type": "string",
        "enum": [
          "PendingDeposit",
          "PendingReview",
          "Confirmed",
          "Cancelled",
          "CheckedIn",
          "InProgress",
          "Completed",
          "Aborted",
          "NoShow"
        ],
        "title": "BookingStatus"
      },
...
      "CancelBookingIn": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "examples": [
              "臨時有事無法前往"
            ]
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "CancelBookingIn",
        "description": "NFR-05 — every cancellation records a reason (retained ≥180 days)."
      },
...
      "CancellationResultOut": {
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/BookingOut"
          },
          "refund": {
            "$ref": "#/components/schemas/RefundOut"
          }
        },
        "type": "object",
        "required": [
          "booking",
          "refund"
        ],
        "title": "CancellationResultOut"
      },
...
      "DepositPaymentIn": {
        "properties": {
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod",
            "examples": [
              "Online"
            ]
          }
        },
        "type": "object",
        "required": [
          "paymentMethod"
        ],
        "title": "DepositPaymentIn",
        "description": "FR-05.3 — deposit = 30% of estimated total."
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Booking rejected: pet danger level is High"
            ]
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "examples": [
              "DANGER_LEVEL_HIGH"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "description": "Shape of error responses (e.g. danger-High rejection, unavailable slot)."
      },
      "FinalPaymentIn": {
        "properties": {
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod",
            "examples": [
              "CardOnSite"
            ]
          }
        },
        "type": "object",
        "required": [
          "paymentMethod"
        ],
        "title": "FinalPaymentIn"
      },
...
      "GroomingType": {
        "type": "string",
        "enum": [
          "Basic",
          "Full"
        ],
        "title": "GroomingType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Money": {
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "examples": [
              "1500.00"
            ]
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "TWD",
            "examples": [
              "TWD"
            ]
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "Money",
        "description": "SDD §3.5: monetary value as decimal + currency. Single-currency system\n(ECPay / Taiwan) so currency defaults to TWD."
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "Online",
          "CardOnSite"
        ],
        "title": "PaymentMethod"
      },
      "PaymentResultOut": {
        "properties": {
          "transactionId": {
            "type": "string",
            "format": "uuid",
            "title": "Transactionid"
          },
          "type": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          }
        },
        "type": "object",
        "required": [
          "transactionId",
          "type",
          "method",
          "amount",
          "status"
        ],
        "title": "PaymentResultOut"
      },
...
      "PaymentType": {
        "type": "string",
        "enum": [
          "Deposit",
          "FinalPay",
          "Refund"
        ],
        "title": "PaymentType"
      },
...
      "ReviewDecision": {
        "type": "string",
        "enum": [
          "Approved",
          "Rejected"
        ],
        "title": "ReviewDecision"
      },
      "ReviewIn": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "staffNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staffnote",
            "examples": [
              "已電話確認，核可"
            ]
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewIn"
      },
...
      "RoomType": {
        "type": "string",
        "enum": [
          "Standard",
          "Deluxe"
        ],
        "title": "RoomType"
      },
      "ServiceCategory": {
        "type": "string",
        "enum": [
          "Lodging",
          "Grooming"
        ],
        "title": "ServiceCategory"
      },
      "ServiceItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/ServiceCategory"
          },
          "basePrice": {
            "$ref": "#/components/schemas/Money"
          },
          "durationMinutes": {
            "type": "integer",
            "title": "Durationminutes"
          },
          "roomType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomType"
              },
              {
                "type": "null"
              }
            ]
          },
          "groomingType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroomingType"
              },
              {
                "type": "null"
              }
            ]
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "basePrice",
          "durationMinutes",
          "isActive"
        ],
        "title": "ServiceItemOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Bearer JWT (enforced in S2)",
        "scheme": "bearer"
      }
    }
  },
...
}